pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logseq-searcher"
version = "0.1.0"
description = "Helps you search your logseq pages and journals"
authors = [{name = "Romilly Cocking", email = "romilly.cocking@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "notebook>=7.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/username/logseq-searcher"
Repository = "https://github.com/username/logseq-searcher.git"
Issues = "https://github.com/username/logseq-searcher/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"